pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG playground"
requires-python = ">=3.10,<3.13"

dependencies = [
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-text-splitters>=0.3.0",
  "langchain-groq>=0.2.0",
  "sentence-transformers>=3.0.0",
  "chromadb>=0.5.0",
  "numpy>=1.26.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "pypdf>=5.0.0",
  "pymupdf>=1.24.0",
  "rich>=13.0.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "ipykernel",
  "pytest",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"